Splitting TripleBuilder class from QueryObjectBuilder
It cuts the class to roughly half and also makes it much more
understandable.
